LayoutTests/imported/w3c/web-platform-tests/html/canvas/element/text/2d.text.draw.baseline.hanging-expected.txt

2d.text.draw.baseline.hanging
Actual output:
Expected output:


FAIL Canvas test: 2d.text.draw.baseline.hanging assert_approx_equals: Red channel of the pixel at (5, 5) expected 0 +/- 2 but got 255
